Split up config into multiple files
There was a lot going on in that one file.

Also, move `GlobalConfig` from `constants.rs` to
`config/global_config.rs`. It's not a constant, even if it has defaults.
It's part of config.
