+++ src/components/memory.rs
@@ -3,7 +3,8 @@ use systemstat::{saturating_sub_bytes, Platform, System};
 use termion::{color, style};
 use thiserror::Error;
 
-use crate::constants::{GlobalConfig, INDENT_WIDTH};
+use crate::constants::INDENT_WIDTH;
+use crate::config::global_config::GlobalConfig;
 
 #[derive(Error, Debug)]
 pub enum MemoryError {
